pextlib1.0: Fix check for SUID/SGID binaries
The S_ISUID and S_ISGID bits are in the st_mode field of the stat
struct, not in the st_flags field. This caused our check for SUID/SGID
binaries to be incorrect, which caused us to copy SUID/SGID binaries
into the sip-workaround directory, where they would then not run
(correctly) because they did not have the permissions required.