darwintrace: Fix path normalization corner cases
There were a few cases in our handling of strcspn(3) that were not
handled correctly, which caused a few corner cases in path normalization
(e.g. walking up past the root directory, multiple slashes in a row) to
not work correctly. Now that we have an automated test suite that covers
this code, those corner cases can be tested and problems identified.